N8N-1 Update continue on fail error handling in fulfillmenttools node
If the continue on fail option is active, then the error message will be returned as part of the `json` property instead of the `error` property. This is based on the implementations in the n8n base nodes. Additionally, the information about the index of the `paired item` for which the error was raised will also be included in the return data.